Create a factory for PlatformProxy
