+++ core/src/CoreTelemetryService.cs
